triagent-autogen/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "triagent-autogen"
version = "0.1.0"
description = "Three-agent AutoGen app with code execution and MCP search via context7"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "autogen-agentchat==0.7.2",
  "autogen-ext[openai,anthropic]==0.7.2",
  "openai>=1.30.0",
  "anthropic>=0.34.0",
  "tiktoken>=0.7.0",
  "typer>=0.12.0",
  "rich>=13.7.1",
  "python-dotenv>=1.0.1",
  "mcp[client]>=1.1.1",
  "websockets>=12.0",
  "uvicorn>=0.30.0",
  "fastapi>=0.111.0",
  "httpx>=0.27.0"
]

[project.scripts]
triagent = "triagent.cli:app"